+++ spring-basic/basic2/exception/summary/섹션9 - API 예외 처리/7.@ExceptionHandler.md
@@ -40,13 +40,10 @@ public class ErrorResult {
 }
 ```
 
-`@ExceptionHandler`에 선언된 `Exception`의 클래스에 따라 에러처리 핸들링이 가능하다.
-
-여기서는 `IllegalArgumentException`을 선언하였기떄문에 에러가나면 IllegalArgumentException에 명시한 메시지를 전달한다.
-
-해당하는 메시지는 미리 만들어준 `ErrorResult`의 필드의 형태를 따라, code와 message 필드에 담아 데이터를 전달한다.
-
-`@ExceptionHandler`는 `@ResponseStatus`를 지정해주지 않으면 자동적으로 HttpStatus.OK(200)의 상태값을 가지며, HttpStatus 지정시 해당하는 상태값에 따라 status를 전달한다.
+`@ExceptionHandler`에 선언된 `Exception`의 클래스에 따라 에러처리 핸들링이 가능하다.<br>
+여기서는 `IllegalArgumentException`을 선언하였기떄문에 에러가나면 IllegalArgumentException에 명시한 메시지를 전달한다.<br>
+해당하는 메시지는 미리 만들어준 `ErrorResult`의 필드의 형태를 따라, code와 message 필드에 담아 데이터를 전달한다.<br>
+`@ExceptionHandler`는 `@ResponseStatus`를 지정해주지 않으면 자동적으로 HttpStatus.OK(200)의 상태값을 가지며, HttpStatus 지정시 해당하는 상태값에 따라 status를 전달한다.<br>
 
 
 ```java
