+++ spring-basic/basic2/exception/summary/섹션9 - API 예외 처리/8.@ControllerAdvice.md
@@ -1,0 +1,33 @@
+controller에 직접 지정한 Exception 메서드들을 ControllerAdvice 클래스에 옮겨와 공통으로 사용하도록 만들어준다.<br>
+ControllerAdvice 클래스에 `@RestControllerAdvice` 붙여 사용하면 전역 및 특정 어노테이션, 패키지, 클래스로 지정하여 예외 처리가 가능하다. 
+
+```java
+@Slf4j
+@RestControllerAdvice                                       // 특정하지 않으면 모든 컨트롤러를 대상으로 ExceptionHandling 을 한다.
+@RestControllerAdvice(annotaions = RestController.class)    // @RestController 이 명시된 controller 에서만 ExceptionHandling 을 한다.
+@RestControllerAdvice("com.practice.exception.api")         // 지정한 패키지내의 Controller에서만 ExceptionHandling 을 한다.
+@RestControllerAdvice(assignedTypes = {ApiV2Controller.class, ApiExceptionController.class})    // 지정한 controller 에서만 ExceptionHandling 을 한다. 
+public class ExControllerAdvice {
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ErrorResult illegalExHandler(IllegalArgumentException e) {
+        log.error("{exceptionHandler} ex", e);
+        return new ErrorResult("BAD", e.getMessage());
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<ErrorResult> userExHandler(UserException e) {
+        log.error("{exceptionHandler} ex", e);
+        ErrorResult errorResult = new ErrorResult("USER-EX", e.getMessage());
+        return new ResponseEntity(errorResult, HttpStatus.BAD_REQUEST);
+    }
+
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler
+    public ErrorResult exHandler(Exception e) {
+        log.error("{exceptionHandler} ex", e);
+        return new ErrorResult("EX", "내부 오류");
+    }
+}
+```
